2020,4,14

+++ src/fly/graphics/FlyCamera.java
@@ -2,9 +2,11 @@ package fly.graphics;
 
 public class FlyCamera
 {
-	public int look_at_x;///<
-	public int look_at_y;///<
+	public double look_at_x;///<
+	public double look_at_y;///<
 
 	public FlyCamera(){this(0,0);}
-	public FlyCamera(int look_at_x,int look_at_y){this.look_at_x = look_at_x;this.look_at_y = look_at_y;}
+
+	public FlyCamera(double look_at_x,double look_at_y)
+	{this.look_at_x = look_at_x;this.look_at_y = look_at_y;}
 }
